Fix CI: PriorityLevel enum size and UDP timeout scaling.
Use uint8_t for PriorityLevel to satisfy clang-tidy performance-enum-size.
Scale UDP TestBase timeout with active test count so the full matrix still
fits CI windows (NUCLEAR_TEST_TIME_UNIT_DEN=10) without restoring the
blanket Windows skip.

+++ src/threading/scheduler/queue/Priority.hpp
@@ -24,6 +24,7 @@
 
 #include <array>
 #include <cstddef>
+#include <cstdint>
 
 namespace NUClear {
 namespace threading {
@@ -31,7 +32,7 @@ namespace threading {
         namespace queue {
 
             /// Fixed scheduler priority buckets (REALTIME, HIGH, NORMAL, LOW, IDLE).
-            enum class PriorityLevel : std::size_t {
+            enum class PriorityLevel : std::uint8_t {
                 REALTIME = 0,
                 HIGH     = 1,
                 NORMAL   = 2,
